Wizard behavior is slightly more intelligent
Wizard moves around in tether and reconsiders if tether is good, even when player is far away. When choosing a new tether, the wizard doesn't pick one that the player is closer to than it is. The energy ball also does a small amount of immediate damage when the player enteres its radius