Seperating Damage and Knockback
For humanoids, we use the exact same calculation used for the player. (a maginal hit) Ground pound knockback now scales inversely with distance (a moderate hit if close). A knockbacked enemy now translates all of its current impact (and direction - this could be tweaked later) to the enemy it collides with. Hopefully this makes comboing enemies easier.