Stopping Spiders From Getting Knockedback
This happens in only one case but it is frequent. When groundpounding spiders and not hitting all of them in damage radius, the farthest will be knocked back. This used to make the spiders go crazy, but now they have such high friction they are unaffected. In the future they should be slightly affected, but not as much as they were before.