Incorporating Knockback Modifier in Everywhere
Now it uses knockback modifier (which it really should have earlier). This doesn't affect gameplay but is good because if we want to make cylinders generally more or less knockbacked, we tweak the one parameter. We already have parameters in place to adjust more specific things (PhysicsMaxForce, PunchPower) so now we can do both.