Rough Implementation of Flying Enemy
Wandering and Dodging are kinda iffy and need to be improved. Right now we can get stuck in walls if we travel exactly on a given axis. Also need to add sound effects, and kill effects. Initial plan was to have it turn into a useless enemy, but I am now less sure about that... midair explosions are fun.